requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --all-extras -o requirements-dev.txt
alembic==1.15.2
    # via flask-migrate
aniso8601==10.0.1
    # via flask-restful
blinker==1.9.0
    # via flask
click==8.1.8
    # via flask
colorama==0.4.6
    # via
    #   click
    #   pytest
flask==3.1.0
    # via
    #   python-api (pyproject.toml)
    #   flask-jwt-extended
    #   flask-migrate
    #   flask-restful
    #   flask-sqlalchemy
flask-jwt-extended==4.7.1
    # via python-api (pyproject.toml)
flask-migrate==4.1.0
    # via python-api (pyproject.toml)
flask-restful==0.3.10
    # via python-api (pyproject.toml)
flask-sqlalchemy==3.1.1
    # via
    #   python-api (pyproject.toml)
    #   flask-migrate
greenlet==3.2.1
    # via sqlalchemy
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
packaging==25.0
    # via pytest
passlib==1.7.4
    # via python-api (pyproject.toml)
pluggy==1.5.0
    # via pytest
psycopg2==2.9.10
    # via python-api (pyproject.toml)
pyjwt==2.10.1
    # via flask-jwt-extended
pytest==8.3.5
    # via python-api (pyproject.toml)
python-dotenv==1.1.0
    # via python-api (pyproject.toml)
pytz==2025.2
    # via flask-restful
six==1.17.0
    # via flask-restful
sqlalchemy==2.0.40
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.13.2
    # via
    #   alembic
    #   sqlalchemy
werkzeug==3.1.3
    # via
    #   flask
    #   flask-jwt-extended
